requirements.txt
aiogram=3.17.0
python-dotenv=1.0.1
matplotlib>=3.5.0
aiohttp>=3.9.3
googletrans==4.0.2
pydantic==2.10.6
